python/worker/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf>=5.0.0
uvloop>=0.19.0; platform_system!="Windows"
torch>=2.3.0
transformers==4.43.3
numpy>=1.24

